db context

+++ Application/Domain/Product.cs
@@ -10,7 +10,7 @@ namespace Application.Domain
     {
         public string Name { get; private set; }
         public string Description { get; private set; }
-        public decimal Price { get; private set; }
+        public double Price { get; private set; }
         public int Stock { get; private set; }
     }
 }
